+++ packages/contracts-rfq/test/FastBridge.t.sol
@@ -937,6 +937,8 @@ contract FastBridgeTest is Test {
         emit BridgeRelayed(
             transactionId, relayer, user, 42161, address(arbUSDC), address(ethUSDC), 11 * 10 ** 6, 10.97e6, 0
         );
+        // Expect not doing any calls to user address
+        vm.expectCall(user, "", 0);
         // Relay the destination bridge
         vm.chainId(1); // set to dest chain
         fastBridge.relay(request);
@@ -990,6 +992,8 @@ contract FastBridgeTest is Test {
             10.97e18,
             0
         );
+        // Expect exactly one call to user address
+        vm.expectCall(user, "", 1);
 
         // Relay the destination bridge
         vm.chainId(1); // set to dest chain
@@ -1046,6 +1050,8 @@ contract FastBridgeTest is Test {
             10.97e18,
             0.005e18
         );
+        // Expect exactly one call to user address
+        vm.expectCall(user, "", 1);
 
         // Relay the destination bridge
         vm.chainId(1); // set to dest chain
@@ -1095,6 +1101,8 @@ contract FastBridgeTest is Test {
         emit BridgeRelayed(
             transactionId, relayer, user, 42161, address(arbUSDC), address(ethUSDC), 11 * 10 ** 6, 10.97e6, 0.005e18
         );
+        // Expect exactly one call to user address
+        vm.expectCall(user, "", 1);
         // Relay the destination bridge
         vm.chainId(1); // set to dest chain
         fastBridge.relay{value: chainGasAmount}(request);
@@ -1103,6 +1111,50 @@ contract FastBridgeTest is Test {
         assertEq(ethUSDC.balanceOf(user), 110.97e6);
         assertEq(relayer.balance, 99.995e18);
         assertEq(user.balance, 0.005e18);
+
+        // Get the returned information of the bridge transaction relays status
+        assertEq(fastBridge.bridgeRelays(transactionId), true);
+
+        // We stop a prank to contain within test
+        vm.stopPrank();
+    }
+
+    function test_successfulRelayDestinationWithChainGasAmountZero() public {
+        // Set up the roles for the test
+        setUpRoles();
+
+        // deal some dest ETH to relayer
+        deal(relayer, 100e18);
+
+        // get bridge request and tx id
+        // chain gas param should be true but we forward 0 amount since not set by governor
+        (bytes memory request, bytes32 transactionId) = _getBridgeRequestAndIdWithChainGas(42161, 0, 0);
+
+        // Get the initial information of the bridge transaction; make sure not relayed
+        assertEq(fastBridge.bridgeRelays(transactionId), false);
+
+        // Start a prank with the relayer
+        vm.startPrank(relayer);
+        // Approve the fastBridge to spend the maximum amount of ethUSDC from the relayer
+        ethUSDC.approve(address(fastBridge), type(uint256).max);
+        // Check the initial balances of the relayer and the user
+        assertEq(ethUSDC.balanceOf(relayer), 100 * 10 ** 6);
+        assertEq(ethUSDC.balanceOf(user), 100 * 10 ** 6);
+        // Expect the BridgeRelayed event to be emitted
+        vm.expectEmit();
+        emit BridgeRelayed(
+            transactionId, relayer, user, 42161, address(arbUSDC), address(ethUSDC), 11 * 10 ** 6, 10.97e6, 0
+        );
+        // Expect not doing any calls to user address
+        vm.expectCall(user, "", 0);
+        // Relay the destination bridge
+        vm.chainId(1); // set to dest chain
+        fastBridge.relay(request);
+        // Check the balances of the relayer and the user after relaying the destination bridge
+        assertEq(ethUSDC.balanceOf(relayer), 89.03e6);
+        assertEq(ethUSDC.balanceOf(user), 110.97e6);
+        assertEq(relayer.balance, 100e18);
+        assertEq(user.balance, 0);
 
         // Get the returned information of the bridge transaction relays status
         assertEq(fastBridge.bridgeRelays(transactionId), true);
@@ -1130,6 +1182,34 @@ contract FastBridgeTest is Test {
         // Relay the destination bridge
         vm.expectRevert(abi.encodeWithSelector(ChainIncorrect.selector));
         vm.chainId(2); // wrong dest chain id
+        fastBridge.relay(request);
+
+        // We stop a prank to contain within test
+        vm.stopPrank();
+    }
+
+    function test_failedRelayTimeExceeded() public {
+        // Set up the roles for the test
+        setUpRoles();
+
+        vm.prank(owner);
+        fastBridge.addRelayer(address(this));
+
+        // get bridge request and tx id
+        (bytes memory request, bytes32 transactionId) = _getBridgeRequestAndId(42161, 0, 0);
+
+        // Start a prank with the relayer
+        vm.startPrank(relayer);
+
+        // Approve the fastBridge to spend the maximum amount of ethUSDC from the relayer
+        ethUSDC.approve(address(fastBridge), type(uint256).max);
+
+        // deadline of 60 min on relay
+        vm.warp(block.timestamp + 61 minutes);
+
+        // Relay the destination bridge
+        vm.expectRevert(abi.encodeWithSelector(DeadlineExceeded.selector));
+        vm.chainId(1); // set to dest chain
         fastBridge.relay(request);
 
         // We stop a prank to contain within test
@@ -1265,7 +1345,8 @@ contract FastBridgeTest is Test {
         // We define a fake transaction hash to be the proof from dest chain
         bytes32 fakeTxnHash = bytes32("0x01");
 
-        vm.warp(block.timestamp + 61 minutes);
+        // user relay deadline is 60 minutes, so add prove period of 60 minutes to that
+        vm.warp(block.timestamp + 121 minutes);
 
         // We provide the relay proof
         vm.expectRevert(abi.encodeWithSelector(DeadlineExceeded.selector));
@@ -1656,7 +1737,8 @@ contract FastBridgeTest is Test {
 
         vm.startPrank(user);
 
-        vm.warp(block.timestamp + 61 minutes);
+        // user relay deadline is 60 minutes, so add prove period of 60 minutes to that
+        vm.warp(block.timestamp + 121 minutes);
 
         vm.expectEmit();
         emit BridgeDepositRefunded(transactionId, user, address(arbUSDC), 11 * 10 ** 6);
@@ -1689,7 +1771,8 @@ contract FastBridgeTest is Test {
 
         vm.startPrank(user);
 
-        vm.warp(block.timestamp + 61 minutes);
+        // user relay deadline is 60 minutes, so add prove period of 60 minutes to that
+        vm.warp(block.timestamp + 121 minutes);
 
         uint256 preRefundBalanceUser = user.balance;
         uint256 preRefundBalanceBridge = address(fastBridge).balance;
@@ -1717,7 +1800,8 @@ contract FastBridgeTest is Test {
 
         vm.startPrank(user);
 
-        vm.warp(block.timestamp + 61 minutes);
+        // user relay deadline is 60 minutes, so add prove period of 60 minutes to that
+        vm.warp(block.timestamp + 121 minutes);
 
         vm.expectEmit();
         emit BridgeDepositRefunded(transactionId, user, address(arbUSDC), 11 * 10 ** 6);
@@ -1751,7 +1835,8 @@ contract FastBridgeTest is Test {
 
         vm.startPrank(user);
 
-        vm.warp(block.timestamp + 61 minutes);
+        // user relay deadline is 60 minutes, so add prove period of 60 minutes to that
+        vm.warp(block.timestamp + 121 minutes);
 
         uint256 preRefundBalanceUser = user.balance;
         uint256 preRefundBalanceBridge = address(fastBridge).balance;
@@ -1778,7 +1863,8 @@ contract FastBridgeTest is Test {
 
         vm.startPrank(user);
 
-        vm.warp(block.timestamp + 61 minutes);
+        // user relay deadline is 60 minutes, so add prove period of 60 minutes to that
+        vm.warp(block.timestamp + 121 minutes);
 
         vm.expectEmit();
         emit BridgeDepositRefunded(transactionId, user, address(arbUSDC), 11 * 10 ** 6);
@@ -1809,7 +1895,8 @@ contract FastBridgeTest is Test {
         // get bridge request and tx id
         (bytes memory request, bytes32 transactionId) = _getBridgeRequestAndId(block.chainid, 0, 0);
 
-        vm.warp(block.timestamp + 61 minutes);
+        // user relay deadline is 60 minutes, so add prove period of 60 minutes to that
+        vm.warp(block.timestamp + 121 minutes);
 
         vm.expectEmit();
         emit BridgeDepositRefunded(transactionId, user, address(arbUSDC), 11 * 10 ** 6);
@@ -1839,7 +1926,7 @@ contract FastBridgeTest is Test {
 
         vm.startPrank(user);
 
-        vm.warp(block.timestamp + 59 minutes);
+        vm.warp(block.timestamp + 119 minutes);
 
         vm.expectRevert(abi.encodeWithSelector(DeadlineNotExceeded.selector));
         fastBridge.refund(request);
@@ -1856,7 +1943,8 @@ contract FastBridgeTest is Test {
         // get bridge request and tx id
         (bytes memory request, bytes32 transactionId) = _getBridgeRequestAndId(block.chainid, 0, 0);
 
-        vm.warp(block.timestamp + 61 minutes);
+        // user relay deadline is 60 minutes, so add prove period of 60 minutes to that
+        vm.warp(block.timestamp + 121 minutes);
 
         vm.expectRevert(abi.encodeWithSelector(StatusIncorrect.selector));
         fastBridge.refund(request);