+++ packages/synapse-interface/components/Banner.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react'
 
-const BANNER_VERSION = '3'
+const BANNER_VERSION = '4'
 
 export const Banner = () => {
   const [hasMounted, setHasMounted] = useState(false)
@@ -47,14 +47,16 @@ export const Banner = () => {
         <div className="m-1 font-thin">
           <div className="container mx-auto">
             <p className="text-md">
-              {/* TODO: replace w/ tweetlink directly */}
+              Arbitrum rebates are live! Distributing 2M $ARB until March 29th -
               <a
+                className="underline text-[#99e6ff] ml-2"
                 href={
-                  'https://twitter.com/SynapseProtocol/status/1726659540551352387'
+                  'https://synapse.mirror.xyz/NpzSkXDUlistuxNQaMwP6HQ9k2gVJsI-G1Y7-gaLxfQ'
                 }
+                rel="noopener noreferrer"
                 target={'_blank'}
               >
-                Synapse now supports Solana at solana.synapseprotocol.com
+                Full details
               </a>
             </p>
           </div>