+++ services/rfq/relayer/quoter/export_test.go
@@ -18,6 +18,10 @@ func (m *Manager) GetQuoteAmount(ctx context.Context, origin, dest int, address 
 	return m.getQuoteAmount(ctx, origin, dest, address, balance)
 }
 
+func (m *Manager) GetDestAmount(ctx context.Context, quoteAmount *big.Int) *big.Int {
+	return m.getDestAmount(ctx, quoteAmount)
+}
+
 func (m *Manager) SetConfig(cfg relconfig.Config) {
 	m.config = cfg
 }